@@ -13,54 +13,54 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^15.1.0",
-		"@angular/cdk": "15.0.4",
-		"@angular/common": "^15.1.0",
-		"@angular/compiler": "^15.1.0",
-		"@angular/core": "^15.1.0",
-		"@angular/forms": "^15.1.0",
-		"@angular/material": "14.2.5",
-		"@angular/platform-browser": "^15.1.0",
-		"@angular/platform-browser-dynamic": "^15.1.0",
-		"@angular/router": "^15.1.0",
-		"@apollo/client": "^3.7.3",
-		"apollo-angular": "4.2.0",
+		"@angular/animations": "^15.1.5",
+		"@angular/cdk": "15.1.5",
+		"@angular/common": "^15.1.5",
+		"@angular/compiler": "^15.1.5",
+		"@angular/core": "^15.1.5",
+		"@angular/forms": "^15.1.5",
+		"@angular/material": "15.1.5",
+		"@angular/platform-browser": "^15.1.5",
+		"@angular/platform-browser-dynamic": "^15.1.5",
+		"@angular/router": "^15.1.5",
+		"@apollo/client": "^3.7.8",
+		"apollo-angular": "4.2.1",
 		"date-fns": "^2.29.3",
 		"graphql": "^16",
-		"highcharts": "^10.3.2",
-		"highcharts-angular": "^3.0.0",
-		"mat-timepicker": "^5.1.7",
+		"highcharts": "^10.3.3",
+		"highcharts-angular": "^3.1.2",
+		"mat-timepicker": "^5.1.8",
 		"rxjs": "~7.8.0",
 		"subscriptions-transport-ws": "^0.11.0",
-		"tslib": "^2.4.1",
+		"tslib": "^2.5.0",
 		"zone.js": "~0.12.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "^15.0.5",
-		"@angular-eslint/builder": "15.1.0",
-		"@angular-eslint/eslint-plugin": "15.1.0",
-		"@angular-eslint/eslint-plugin-template": "15.1.0",
-		"@angular-eslint/schematics": "15.1.0",
-		"@angular-eslint/template-parser": "15.1.0",
-		"@angular/cli": "~15.0.5",
-		"@angular/compiler-cli": "^15.1.0",
-		"@graphql-codegen/cli": "^2.16.4",
-		"@graphql-codegen/typescript": "^2.8.7",
+		"@angular-devkit/build-angular": "^15.1.6",
+		"@angular-eslint/builder": "15.2.1",
+		"@angular-eslint/eslint-plugin": "15.2.1",
+		"@angular-eslint/eslint-plugin-template": "15.2.1",
+		"@angular-eslint/schematics": "15.2.1",
+		"@angular-eslint/template-parser": "15.2.1",
+		"@angular/cli": "~15.1.6",
+		"@angular/compiler-cli": "^15.1.5",
+		"@graphql-codegen/cli": "^3.0.0",
+		"@graphql-codegen/typescript": "^3.0.0",
 		"@graphql-codegen/typescript-apollo-angular": "^3.5.6",
-		"@graphql-codegen/typescript-operations": "^2.5.12",
-		"@types/jest": "^29.2.5",
-		"@typescript-eslint/eslint-plugin": "5.48.1",
-		"@typescript-eslint/parser": "5.48.1",
+		"@graphql-codegen/typescript-operations": "^3.0.0",
+		"@types/jest": "^29.4.0",
+		"@typescript-eslint/eslint-plugin": "5.52.0",
+		"@typescript-eslint/parser": "5.52.0",
 		"autoprefixer": "^10.4.13",
-		"eslint": "^8.31.0",
+		"eslint": "^8.34.0",
 		"eslint-config-prettier": "^8.6.0",
 		"eslint-plugin-prettier": "^4.2.1",
-		"jest": "^29.3.1",
-		"jest-preset-angular": "^12.2.3",
+		"jest": "^29.4.3",
+		"jest-preset-angular": "^13.0.0",
 		"karma-coverage": "~2.2.0",
 		"postcss": "^8.4.21",
 		"prettier-eslint": "^15.0.1",
-		"tailwindcss": "^3.2.4",
+		"tailwindcss": "^3.2.7",
 		"ts-node": "^10.9.1",
 		"typescript": "~4.8.4"
 	}
